requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1

# Machine Learning
torch
torchaudio
transformers==4.37.2
numpy
scikit-learn==1.3.2

# File type detection
python-magic==0.4.27

# HTTP client (for testing)
httpx==0.26.0
requests==2.31.0

# Utilities
pydub==0.25.1